Add CNROM & NROM files
