+++ src/mappers/CNROM.h
@@ -1,0 +1,9 @@
+#ifndef NESULATOR_CNROM_H
+#define NESULATOR_CNROM_H
+
+#include "../rom.h"
+#include "../memory.h"
+
+void mapper3(struct ROM *rom);
+
+#endif //NESULATOR_CNROM_H
