+++ src/mappers/NROM.c
@@ -1,0 +1,5 @@
+#include "NROM.h"
+
+void mapper0(struct ROM *rom) {
+
+}
