+++ src/mappers/NROM.h
@@ -1,0 +1,8 @@
+#ifndef NESULATOR_NROM_H
+#define NESULATOR_NROM_H
+
+#include "../rom.h"
+
+void mapper0(struct ROM *rom);
+
+#endif //NESULATOR_NROM_H
